+++ release_notes.md
@@ -28,3 +28,6 @@
    message_id -> message_ids.
 3. Renamed GetMessages to GetMessagesByBookByStreams.
 4. [TH2-4581] Added a check that commands start & end time parameters are milliseconds.
+5. [TH2-4625] Commands added:
+   - GetPage command added to get Page by name with book_id
+   - GetPageAll command added to get all Pages with book_id